refactor: rename rate_limiter to sec_rate_limiter for clarity
  Rename SEC-specific rate limiter for better clarity:
  - Move from rate_limiter.py to sec_rate_limiter.py
  - Update all test imports to use new module name
  - Maintain same functionality for SEC Edgar API compliance
  - Separate from new API client rate limiting functionality
